requirements.txt
transformers
benepar
evaluate
inflection
matplotlib
nltk
numpy
pandas
sacrebleu
scikit-learn
scipy
seaborn
spacy
tqdm
unbabel-comet
wandb
spacy-lefff
inflecteur
sacremoses
lemminflect